Create the price of laptops
The specifications of the laptop will be given . The first line of the input contains the number n, which indicates the number of laptops. Each of the next n lines contains two numbers, the first number shows the price of the laptop and the second number shows the quality of that laptop. If the program find two laptops that meets the conditions mentioned by user, it will print, happy users' name, otherwise, prints poor users' name.