Create World cup group B table.py
The teams in Group B of the World Cup are Iran, Portugal, Spain, and Morocco. Write a program that, upon receiving the results of the games, prints the team name, number of wins and losses, goal difference, and points on one line. Each team should be printed in order of points on one line. (If the points were equal, the number of wins should be considered. If both the number of wins and points were equal, they should be printed in alphabetical order.)

Note: A team gets zero points for a loss, one point for a draw, and three points for a win.
Goal difference is the difference between goals scored and goals conceded by a team.

Read the results of the games in the following order: (In the sample input, the number on the left corresponds to the team on the right.)
Iran – Spain
Iran – Portugal
Iran – Morocco
Spain – Portugal
Spain – Morocco