Create Favorite Genres survay.py
In a survey of people interested in watching movies, they were asked to write down their three favorite genres. They were given six different genres to choose from, including:

Horror, Romance, Comedy, History, Adventure, Action

Write a program that takes the number of people, then takes the name of each person with the genres they are interested in, and prints the name of each genre and the number of people interested in that genre in order of greatest interest in the output. (If the interest in different genres is the same, print it in alphabetical order in the output.) If a genre is not selected, consider its value as zero and print the name and the number 0 in the output.