Create Key words.py
Write a program that prints the index words (words that start with a capital letter) from a text along with the word number (the most frequent word). If no word with this attribute is found in the text, print None in the output. Words that are at the beginning of a sentence should not be considered as index words. (Start the word number from one)

Numbers are not considered index words. The only punctuation mark used in a sentence other than a period is a comma. Be sure to remove any period or comma at the end of a word.