Create Health program.py
Two different health programs are planned in two schools, A and B. Milk is distributed in one school and not in the other. Nutrition experts plan to compare the height, weight, and age of students in these two classes by examining their ages. Write a program that takes the number of students in each class, plus their ages, heights, and weights, stores each in a list, and prints the average age, heights, and weights of each class on a separate line in the output (printed as floats), and then prints the class with the highest average height in the output. (First, enter the number and information of class A, and then the information of class B. The output should also be printed in the order of the input information.) If the average height is equal, the class with the lowest average weight should be printed. If the average height and weight are equal, the statement Same (exactly the same) should be printed in the output.