Create Current age.py
Write a program that uses the datetime library class in Python to get the age of a person. If the input date is incorrect, print the word WRONG in the output. The date format is yyyy/mm/dd. (For example, if the month is entered as 15, because we have only 12 months, it should print WRONG in the output. The same applies to the number of days, and the maximum is 31.)

Note: The input and output of the sample below is from running the program on 2019/02/01.