Create Employee information.py
A database containing the names, heights, and weights of employees of a company is designed. Write a program that reads the information from this database file and prints the employees in order of height in the output along with their names and weights. (If they are of equal height, print the person who weighs less higher.) It is guaranteed that no two people are the same height or weight.

The format of the database table information is as follows:

Sample Input: (The following information is defined in the database and this is an example and the number of people can vary)