Create Saving login information.py
Write a program that takes email and password as input and stores them in a database as username and password. Submit only your code file. The code written should be able to take the information as input and store and log the information in a desired database file.

It is essential to follow the email format as input. If the input format is not in the email format, show the user the correct example and ask the user for input again. The password also includes numbers and strings.

The correct email pattern is one that consists of two strings and an expression (including a string or a number and a string at the same time) separated by @ and ., as the desired correct email. As shown below:

expression@string.string