A project .py
Get all the country information from Scrapethissite and save it in a database of your choice. Then try to estimate the area of ​​the country using the population of the country.
Note: Due to the lack of data, the prediction may not be accurate, so only the steps of implementing the model are considered for prediction.